requirements.txt
streamlit
gtts
speechrecognition
pyaudio
deep-translator
langdetect
google-generativeai>=0.4.0
requests>=2.31.0
numpy
pandas
#langchain 
#wikipedia 
#langchain-community 
#langchain-openai
#langchain-anthropic
#python-dotenv
#pydantic
#duckduckgo-search